Changed sceneLabel2.py example to use ultraflow library.
ultraflow is in the ./maxflow subdir, it's our interface to the B&K
code, replacing PyMaxflow.

When you run the demo you have to tell it where to find our module,
using the PYTHONPATH variable.  In bash you can use "export PYTHONPATH=..."
to set it permanently (or in your .bashrc), or on the command-line
when you run it like this:

> PYTHONPATH=./maxflow ./sceneLabel2.py